// src/ts/common/interfaces.ts
export const enum NotificationFlags {
	None = 0,
	Ok = 1,
	Yes = 2,
	No = 4,
	Accept = 8,
	Reject = 16,
	Supporter = 32,
	Ignore = 64,
	NameBad = 128,
}
// ...
export const enum FriendStatusFlags {
	None = 0,
	Online = 1,
	Remove = 2,
}

export interface FriendStatusData {
	accountId: string;
	accountName?: string;
	status: FriendStatusFlags;
	entityId?: number;
	name?: string;
	info?: string;
	crc?: number;
	nameBad?: boolean;
}